move github mock utility to test package
